# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BlueMoxon API",
    "description": "Victorian Book Collection Management API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/publishers": {
      "get": {
        "tags": [
          "publishers"
        ],
        "summary": "List Publishers",
        "description": "List all publishers.",
        "operationId": "list_publishers_api_v1_publishers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "publishers"
        ],
        "summary": "Create Publisher",
        "description": "Create a new publisher.",
        "operationId": "create_publisher_api_v1_publishers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublisherCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublisherResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/publishers/{publisher_id}": {
      "get": {
        "tags": [
          "publishers"
        ],
        "summary": "Get Publisher",
        "description": "Get a single publisher with their books.",
        "operationId": "get_publisher_api_v1_publishers__publisher_id__get",
        "parameters": [
          {
            "name": "publisher_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Publisher Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "publishers"
        ],
        "summary": "Update Publisher",
        "description": "Update a publisher.",
        "operationId": "update_publisher_api_v1_publishers__publisher_id__put",
        "parameters": [
          {
            "name": "publisher_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Publisher Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublisherUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublisherResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "publishers"
        ],
        "summary": "Delete Publisher",
        "description": "Delete a publisher. Will fail if publisher has associated books.",
        "operationId": "delete_publisher_api_v1_publishers__publisher_id__delete",
        "parameters": [
          {
            "name": "publisher_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Publisher Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/export/csv": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Export Csv",
        "description": "Export books to CSV format matching PRIMARY_COLLECTION.csv structure.",
        "operationId": "export_csv_api_v1_export_csv_get",
        "parameters": [
          {
            "name": "inventory_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "PRIMARY",
              "title": "Inventory Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/export/json": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Export Json",
        "description": "Export books to JSON format with all details.",
        "operationId": "export_json_api_v1_export_json_get",
        "parameters": [
          {
            "name": "inventory_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "PRIMARY",
              "title": "Inventory Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PublisherCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier"
          },
          "founded_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Founded Year"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PublisherCreate",
        "description": "Schema for creating a publisher."
      },
      "PublisherResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier"
          },
          "founded_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Founded Year"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "book_count": {
            "type": "integer",
            "title": "Book Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "PublisherResponse",
        "description": "Publisher response schema."
      },
...
      "PublisherUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier"
          },
          "founded_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Founded Year"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "PublisherUpdate",
        "description": "Schema for updating a publisher."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}